Server-side python SDK added __init__.py file
Summary: The __init__.py files are required to make Python treat directories containing the file as packages. This prevents directories with a common name, such as string, unintentionally hiding valid modules that occur later on the module search path. In the simplest case, __init__.py can just be an empty file

Reviewed By: TobiasWooldridge

Differential Revision: D18812905

fbshipit-source-id: 92ee7d2f