Fix Cursor Paging Issue on Duplicating Fields
Summary:
[BizSDK][Python] Fix Cursor Paging Issue on Duplicating Fields

https://github.com/facebook/facebook-python-business-sdk/issues/641

We have seen access_token and summary fields being duplicated during multiple API calls using the Cursor on paging. A newly placed (?) restriction on the size of the request surfaced this issue up because the requests getting larger and larger when calling load_next_page() function again and again.

To check the details of this bug, please refer to the issue reported by the community.

After the fix, the version will also be bumped to 17.0.2.

Reviewed By: mengxuanzhangz

Differential Revision: D46949894

fbshipit-source-id: 2fdbb49f309b4f7709a4bdccdf9e80e53355268a