Fix Python Paging Infinite Issue
Summary: This diff fixes a bug in the Python SDK for Facebook Business where the paging cursor was not properly checked for the existence of the 'next' key, which caused an infinite loop when there were no more pages to fetch. The code changes in the 'api.py' file of the SDK and the 'facebook_business/api.py' file of the codegen templates ensure that the 'after' cursor is only set if the 'next' key exists in the paging information. This should prevent the infinite loop and allow the SDK to properly handle paging in the API calls.

Reviewed By: mengxuanzhangz

Differential Revision: D53010767

fbshipit-source-id: 6588814b8fb48be213bedc4b7e304aeecdca0211